New implementation, better performance
The exact 1 to 1 implementation from MATLAB I tried but the performance was very bad some of the stuff from MATLAB doesn't replicate exactly in Python, so I improved it with KD-trees and a bit of vectorized code-keeping the logic the same. But, still, I got the same results as the MATLAB implementation at threshold 1.25 and not at 1.0. (Not sure exactly what's wrong this time)